Implement deletion of customer profile when customer deleted

+++ delete.php
@@ -2,14 +2,14 @@
 <?php include "includes/functions.php"; ?>
 <?php
 
-debug_to_console("Inside delete.php");
-
 if(isset($_POST['cust_id'])){
 
     // Delete Customer    
     $cust_id = clean($_POST['cust_id']);
     if(delete_addresses_by_cust($cust_id)){
-        delete_cust($cust_id);
+        if(delete_profile_pic($cust_id)){
+            delete_cust($cust_id);
+        }
     }
 }
